Refactor Wall: Remove the move method in Wall Class
The move method includes both Player and Ball’s move method which is not necessary, because they can be called separately.

And with the move method in Wall class, the code used is [ wall.move() ] instead of [ ball.move() and player.move() ] which is confusing for the code reader, because the reader doesn’t know what objects are actually moving with that code.

Hence, I decided to remove it to increase the readability of the code.