Refactor Wall: Create randomSpeedX and randomSpeedY methods in Wall class
These methods are created for randomly setting the speed of the ball.

I decided to create them so that they can be reused in multiple places whenever there is a need to set the speed of the ball instead of writing lines of code to do it every time like last time.