Refactor Wall: Remove ballEnd method from Wall class
This is the third step of clearing out all the Ball’s related responsibilities from Wall class.

The ballEnd method is used to return true if ball count is equal to 0.

I decided to remove it because its name can be confusing, because it could mean the ball is lost, and I prefer to use (ball.getCount == 0) directly which the meaning is clear.