Refactor Wall: Create BrickCount class and move getBrickCount, isDone, brickCountDecrement, setBrickCount methods from Wall class to it
This is my last approach to make Wall class have only single responsibility by moving our brick count’s related methods from it.

So I decided to create a new class BrickCount to handle all the logics that are related to brick count.

And now the Wall class is clean and only responsible for creating objects for the relevant game components.