Refactor HomeMenu to follow MVC pattern: Create new classes HomeMenuView and HomeMenuController
This is my first step to refactor HomeMenu to follow the MVC design pattern.

HomeMenuView will be responsible for creating the start button and exit button objects to interact with users.

HomeMenuController will be responsible for creating methods for the buttons to perform actions like starting the game and quitting the game.

I haven’t created a model for the home menu because there is no data/state that needs to be updated in this view at this moment.