Refactor DebugConsole to follow MVC pattern: Create new classes DebugConsoleView and DebugConsoleController
This is my first step to refactor DebugConsole to follow the MVC design pattern.

DebugConsoleView will be responsible for managing and defining how the buttons and sliders are presented to the user.

DebugConsoleController will be responsible for creating logic for the buttons and sliders to perform action to interact with relevant models like ball and bricks that are inside of Wall class at this moment.